Refactored tgtdb package code to use testcontainers library
