+++ yb-voyager/src/srcdb/main_test.go
@@ -62,6 +62,7 @@ func TestMain(m *testing.M) {
 			User:      postgresContainer.GetConfig().User,
 			Password:  postgresContainer.GetConfig().Password,
 			Schema:    postgresContainer.GetConfig().Schema,
+			DBName: postgresContainer.GetConfig().DBName,
 			Host:      host,
 			Port:      port,
 			SSLMode:   "disable",
