+++ yb-voyager/test/containers/test_schemas/postgresql_schema.sql
@@ -4,11 +4,13 @@ CREATE TABLE public.foo (
         id INT PRIMARY KEY,
         name VARCHAR
 );
+INSERT into public.foo values (1, 'abc'), (2, 'xyz');
 
 CREATE TABLE public.bar (
         id INT PRIMARY KEY,
         name VARCHAR
 );
+INSERT into public.bar values (1, 'abc'), (2, 'xyz');
 
 CREATE TABLE public.unique_table (
         id SERIAL PRIMARY KEY,
@@ -31,11 +33,11 @@ CREATE TABLE public.table2 (
 );
 
 CREATE TABLE public.non_pk1(
-    id INT,
-    name VARCHAR(255)
+        id INT,
+        name VARCHAR(255)
 );
 
 CREATE TABLE public.non_pk2(
-    id INT,
-    name VARCHAR(255)
+        id INT,
+        name VARCHAR(255)
 );
